+++ nextstep-di/src/main/java/nextstep/di/factory/BeanFactory.java
@@ -3,6 +3,7 @@ package nextstep.di.factory;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
 import nextstep.annotation.Inject;
+import nextstep.di.factory.exception.*;
 import nextstep.stereotype.Controller;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
