+++ nextstep-di/src/test/java/nextstep/di/factory/BeanFactoryTest.java
@@ -2,6 +2,7 @@ package nextstep.di.factory;
 
 import com.google.common.collect.Sets;
 import nextstep.di.factory.example.*;
+import nextstep.di.factory.exception.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
